split tests file hierarchy
